+++ src/ray.h
@@ -8,21 +8,21 @@ class Ray
 private:
   Vec3 origin;
   Vec3 dir;
-  double time_;
+  float time_;
 
 public:
   __host__ __device__ Ray() {};
-  __host__ __device__ Ray(const Vec3 &ori, const Vec3 &di, double time) : origin(ori), dir(di), time_(time) {};
-  __host__ __device__ Ray(const Vec3 &ori, const Vec3 &di) : origin(ori), dir(di), time_(0.0) {};
+  __host__ __device__ Ray(const Vec3 &ori, const Vec3 &di, float time) : origin(ori), dir(di), time_(time) {};
+  __host__ __device__ Ray(const Vec3 &ori, const Vec3 &di) : origin(ori), dir(di), time_(0.0f) {};
   __host__ __device__ const Vec3 &get_origin() const { return origin; }
   __host__ __device__ const Vec3 &get_direction() const { return dir; }
 
   // Calculate ray position with given time t
-  __host__ __device__ Vec3 at(double t) const
+  __host__ __device__ Vec3 at(float t) const
   {
     return origin + t * dir;
   }
-  __host__ __device__ double get_time() const
+  __host__ __device__ float get_time() const
   {
     return time_;
   }
